Добавлена поддержка дробных чисел.

+++ UtilityBot/Services/ITextHandler.cs
@@ -7,6 +7,6 @@ namespace UtilityBot.Services
     public interface ITextHandler
     {
         string[] MessageParse(Message message, string userFunction, out bool error);
-        int Process(string[] massive, string userFunction);
+        decimal Process(string[] massive, string userFunction);
     }
 }
